working at root dir
