+++ src/setup/index.php
@@ -16,6 +16,7 @@ require_once __DIR__ . '/../vendor/autoload.php';
 // For /churchcrm/setup/index.php -> SCRIPT_NAME = /churchcrm/setup/index.php
 // For /setup/index.php -> SCRIPT_NAME = /setup/index.php
 $scriptName = $_SERVER['SCRIPT_NAME'] ?? '/setup/index.php';
+$scriptName = str_replace('\\', '/', $scriptName);
 $basePath = dirname($scriptName); // Gets /churchcrm/setup or /setup
 
 // Calculate root path (parent of setup directory)
@@ -24,6 +25,28 @@ $basePath = dirname($scriptName); // Gets /churchcrm/setup or /setup
 $rootPath = dirname($basePath);
 if ($rootPath === '/' || $rootPath === '.') {
     $rootPath = '';
+}
+
+// Fallback detection when SCRIPT_NAME lacks the installation prefix (common with subdir installs)
+if ($rootPath === '') {
+    $requestUri = $_SERVER['REQUEST_URI'] ?? '';
+    $requestPath = parse_url($requestUri, PHP_URL_PATH) ?? '';
+    if ($requestPath !== '') {
+        $requestPath = rtrim($requestPath, '/');
+        if ($requestPath !== '') {
+            $candidate = preg_replace('#/setup(?:/.*)?$#', '', $requestPath);
+            if ($candidate !== null && $candidate !== $requestPath) {
+                $candidate = rtrim($candidate, '/');
+                if ($candidate !== '') {
+                    $rootPath = $candidate;
+                }
+            }
+        }
+    }
+}
+
+if ($rootPath !== '' && $rootPath[0] !== '/') {
+    $rootPath = '/' . $rootPath;
 }
 
 // Store paths in global for template access (no SystemURLs available)
